+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 Release History
 ---------------
 ## [Unreleased]
+
+# [1.1.0] - 2017-01-19
+
+### Breaking change
+- Selector changed from `AddExample` to `exemplify`
+
+### Added
+- Option for passing angular inputs (see issue #1)
+- Option for passing custom texts
+
+### Fixed
+- Layout when using bootstrap 4 alpha 6
+
 # [1.0.0] - 2017-01-03
 
 ### Added